requirements.txt
accelerate==1.4.0
datasets==3.2.0
fire==0.7.0
transformers==4.50.0
torch==2.6.0
pillow==11.1.0
deepspeed
tqdm
nltk
wandb